% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBPE_Dunn.R
\name{plotBPE_Dunn}
\alias{plotBPE_Dunn}
\title{A function to draw bar (B) plot with point (P) and error (E) bars}
\usage{
plotBPE_Dunn(
  dat,
  levels,
  comparison,
  style = 2,
  plot.B = T,
  lab.x = "",
  lab.y = "",
  title = "",
  text.x,
  size.text.x = 10,
  size.text.y = 10,
  size.point = 2,
  size.p = 3,
  angle.x = 0,
  breaks.y = NULL,
  fill,
  jitter.width = 0.1,
  aspect.ratio = 0.5
)
}
\arguments{
\item{dat}{a data frame}

\item{levels}{level 1 must be the control group}

\item{comparison}{a numeric vector indicates which groups are used to compare with control}

\item{style}{1 or 2}

\item{plot.B}{boolean values determining if bar plot should be drawn}

\item{lab.x}{label of x-axis}

\item{lab.y}{label of y-axis}

\item{title}{title}

\item{text.x}{x-axis text}

\item{size.text.x}{size of x-axis text}

\item{size.text.y}{size of y-axis text}

\item{size.point}{size of point}

\item{size.p}{size of p values}

\item{angle.x}{angle of x-axis text}

\item{breaks.y}{breaks of y-axis}

\item{fill}{fill colors}

\item{jitter.width}{jitter width of the points}

\item{aspect.ratio}{ratio of the x axis and y axis}
}
\value{
a ggplot2 plot
}
\description{
A function to draw bar (B) plot with point (P) and error (E) bars
}
