% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBPE_T.R
\name{plotBPE_T}
\alias{plotBPE_T}
\title{Title}
\usage{
plotBPE_T(
  dat,
  exp = "exp",
  group = "group",
  levels,
  comparison,
  style = 1,
  plot.B = T,
  lab.x = "",
  lab.y = "",
  title = "",
  text.x,
  size.text.x = 10,
  size.text.y = 10,
  size.point = 2,
  size.p = 4,
  angle.x = 0,
  breaks.y = NULL,
  start.y = 0,
  fill = "white",
  jitter.width = 0.1,
  aspect.ratio = 0.5
)
}
\arguments{
\item{aspect.ratio}{}
}
\description{
Title
}
