% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFC_1group.R
\name{plotFC_1group}
\alias{plotFC_1group}
\title{Density plot for flow cytometry}
\usage{
plotFC_1group(
  file,
  label,
  channel,
  gate = c(1, 7),
  xlims = c(0.5, 8),
  text.x = 7,
  lab.x = NULL,
  lab.y = "\% of Max",
  title = "xxx",
  legend = "legend",
  fill = "#939393",
  color = "#939393",
  alpha = 0.5,
  fixed_ratio = 5.5
)
}
\arguments{
\item{file}{a standard CSV file generated by NovoExpress software}

\item{label}{label of sample}

\item{channel}{channel name in the CSV file}

\item{gate}{threshold}

\item{xlims}{two numeric values, specifying the lower limit and the upper limit of the scale}

\item{text.x}{text}

\item{lab.x}{x-axis label}

\item{lab.y}{y-axis label}

\item{title}{plot title}

\item{legend}{plot legend, the vaules can be "none"/"legend"}

\item{fill}{fill}

\item{color}{color}

\item{alpha}{alpha}

\item{fixed_ratio}{ratio of the x axis and y axis}
}
\description{
Density plot for flow cytometry
}
