% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFC_2d.R
\name{plotFC_2d}
\alias{plotFC_2d}
\title{2D plot for flow cytometry}
\usage{
plotFC_2d(
  file,
  label = "none",
  channel,
  xlims = c(3, 8),
  ylims = c(2, 7),
  gates = c(4.1, 3.5),
  legend = "none",
  type = c("point", "density", "point_density"),
  alpha.point = 0.3,
  alpha.density = 0.6,
  size.point = 0.1,
  size.density = 1,
  ajust.pointdensity = 0.5,
  bins.density = 50
)
}
\arguments{
\item{file}{a standard CSV file generated by NovoExpress software}

\item{label}{label of sample}

\item{channel}{channel name in the CSV file}

\item{xlims}{two numeric values, specifying the lower limit and the upper limit of the scale}

\item{ylims}{two numeric values, specifying the lower limit and the upper limit of the scale}

\item{gates}{two numeric values, specifying the gates}

\item{legend}{plot legend, the values can be "none"/"legend"}

\item{type}{plot type, the values can be "point"/"density"}

\item{alpha.point}{alpha of the point}

\item{alpha.density}{alpha of the density}

\item{size.point}{size of the point}

\item{size.density}{size of the density}

\item{bins.density}{bins of the density}

\item{adjust.pointdensity}{adjust for pointdensity plot}
}
\value{
a ggplot2 plot
}
\description{
2D plot for flow cytometry
}
