% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFC_2group.R
\name{plotFC_2group}
\alias{plotFC_2group}
\title{Density plot for flow cytometry}
\usage{
plotFC_2group(
  file_con,
  file_test,
  label_con = "con",
  label_test = "test",
  channel,
  gate,
  xlims,
  text.x,
  lab.x = NULL,
  lab.y = "\% of Max",
  title = "xxx",
  legend = "legend",
  fills,
  colors,
  alpha,
  fixed_ratio = 5.5
)
}
\arguments{
\item{file_con}{a standard CSV file of con sample generated by NovoExpress software}

\item{file_test}{a standard CSV file of test sample generated by NovoExpress software}

\item{label_con}{label of con sample}

\item{label_test}{label of test sample}

\item{channel}{channel name in the CSV file}

\item{gate}{high threshold, use 7 to indicate 10^7, the low threshold is automatically detected by the 0.99 quantile of the con sample}

\item{xlims}{two numeric values, specifying the lower limit and the upper limit of the scale}

\item{text.x}{text}

\item{lab.x}{x-axis label}

\item{lab.y}{y-axis label}

\item{title}{plot title}

\item{legend}{plot legend, the vaules can be "none"/"legend"}

\item{fills}{colors used to fill the density}

\item{colors}{colors}

\item{alpha}{alpha}

\item{fixed_ratio}{ratio of the x axis and y axis}
}
\value{
a ggplot2 density plot
}
\description{
The function is designed to generate a two-sample density plot for flow cytometry assay
}
\examples{
# none
}
