% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFC_multi.R
\name{plotFC_multi}
\alias{plotFC_multi}
\title{Density plot for flow cytometry}
\usage{
plotFC_multi(
  files,
  labels,
  pos.label = 0.3,
  nudge.label.y = 0.8,
  plot.perc = F,
  quantile.probs = 0.99,
  nudge.perc.x = 3,
  nudge.perc.y = 0.8,
  plot.MFI = F,
  nudge.MFI.x = 3,
  nudge.MFI.y = 0.8,
  channel,
  xlims,
  title = "xxx",
  fills,
  colors,
  alpha,
  fixed_ratio = 3
)
}
\arguments{
\item{files}{standard CSV files generated by NovoExpress software}

\item{labels}{labels}

\item{pos.label}{position of labels}

\item{nudge.label.y}{Vertical adjustment to nudge sample labels by}

\item{plot.perc}{logical, it controls whether to plot perc labels}

\item{quantile.probs}{numeric vector of probabilities with values in 0-1}

\item{nudge.perc.x}{horizontal adjustment to nudge perc labels by}

\item{nudge.perc.y}{vertical adjustment to nudge perc labels by}

\item{plot.MFI}{logical, it controls whether to plot MFI labels}

\item{nudge.MFI.x}{horizontal adjustment to nudge MFI labels by}

\item{nudge.MFI.y}{vertical adjustment to nudge MFI labels by}

\item{channel}{channel name in the CSV file}

\item{xlims}{two numeric values, specifying the lower limit and the upper limit of the scale}

\item{title}{plot title}

\item{fills}{colors used to fill the density}

\item{colors}{colors}

\item{alpha}{alpha}

\item{fixed_ratio}{ratio of the x axis and y axis}
}
\value{
a ggplot2 density plot
}
\description{
The function is designed to generate a multi-sample density plot for flow cytometry assay
}
\examples{
# none

}
