% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ_t.R
\name{plotQ_t}
\alias{plotQ_t}
\title{Gene expression plot for qPCR analysis}
\usage{
plotQ_t(
  file,
  genes,
  groups,
  title,
  fill,
  label.group,
  size.p = 2,
  angle.p = 45
)
}
\arguments{
\item{file}{a table file of gene expression generated by QuantStudio software}

\item{genes}{genes to plot}

\item{groups}{two groups}

\item{title}{plot title}

\item{fill}{colors}

\item{label.group}{group labels}

\item{size.p}{font size for p value}

\item{angle.p}{angle for p value}
}
\value{
a ggplot2 bar plot
}
\description{
Gene expression plot for qPCR analysis
}
