% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ_Tukey.R
\name{plotQ_Tukey}
\alias{plotQ_Tukey}
\title{Gene expression plot for qPCR analysis}
\usage{
plotQ_Tukey(file, gene, geneLabel, fill, groups, comparison, style)
}
\arguments{
\item{file}{a table file of gene expression generated by QuantStudio software}

\item{gene}{genes to plot}

\item{geneLabel}{labels of genes}

\item{fill}{colors}

\item{groups}{groups}

\item{comparison}{comparison}

\item{style}{1 or 2}
}
\description{
Gene expression plot for qPCR analysis
}
