% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq_calcu.R
\name{rq_calcu}
\alias{rq_calcu}
\title{Calculate relative expression from Ct values}
\usage{
rq_calcu(
  exp,
  group = "group",
  gene = "gene",
  ct = "ct",
  gene_ctrl,
  sample_ctrl
)
}
\arguments{
\item{exp}{a data frame}

\item{group}{group column}

\item{gene}{gene column}

\item{ct}{ct value column}

\item{gene_ctrl}{contrl gene}

\item{sample_ctrl}{contrl group}
}
\value{
a data frame of exp
}
\description{
Calculate relative expression from Ct values
}
\examples{
none
}
