% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciBPE.R
\name{sciBPE}
\alias{sciBPE}
\title{A function to draw bar (B) plot with point (P) and error (E) bars}
\usage{
sciBPE(
  dat,
  plot_B = T,
  levels,
  lab.sample = NULL,
  lab.x = "",
  lab.y = "",
  size.text.x = 10,
  size.text.y = 10,
  size.point = 2,
  title = "",
  fill = "black",
  ylims,
  aspect.ratio = 0.5
)
}
\arguments{
\item{dat}{a data frame}

\item{plot_B}{boolean values determining if bar plot should be drawn}

\item{levels}{levels of samples}

\item{lab.sample}{labels of samples}

\item{lab.x}{label of x-axis}

\item{lab.y}{label of y-axis}

\item{size.text.x}{size of x-axis text}

\item{size.text.y}{size of y-axis text}

\item{size.point}{size of point}

\item{title}{title}

\item{fill}{fill colors}

\item{ylims}{two numeric values, specifying the lower limit and the upper limit of the scale}

\item{aspect.ratio}{ratio of the x axis and y axis}
}
\value{
a ggplot2 plot
}
\description{
A function to draw bar (B) plot with point (P) and error (E) bars
}
