% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBPE_T.R
\name{plotBPE_T}
\alias{plotBPE_T}
\title{A function to draw bar (B) plot with point (P) and error (E) bars with t-test}
\usage{
plotBPE_T(
  dat,
  formula,
  comparisons,
  plot.B = T,
  lab.x = "",
  lab.y = "",
  title = "",
  text.x,
  size.text.x = 10,
  size.text.y = 10,
  size.point = 2,
  size.p = 4,
  nudge.p = 1.1,
  angle.x = 0,
  breaks.y = NULL,
  start.y = 0,
  fill = "white",
  aspect.ratio = 0.5
)
}
\arguments{
\item{dat}{a data frame}

\item{formula}{a formula}

\item{comparisons}{A list of length-2 vectors specifying the groups of interest to be compared}

\item{plot.B}{boolean values determining if bar plot should be drawn}

\item{lab.x}{label of x-axis}

\item{lab.y}{label of y-axis}

\item{title}{title}

\item{text.x}{x-axis text}

\item{size.text.x}{size of x-axis text}

\item{size.text.y}{size of y-axis text}

\item{size.point}{size of point}

\item{size.p}{size of p values}

\item{nudge.p}{nudge of p values}

\item{angle.x}{angle of x-axis text}

\item{breaks.y}{breaks of y-axis}

\item{start.y}{start of y-axis}

\item{fill}{fill colors}

\item{aspect.ratio}{ratio of the x axis and y axis}
}
\value{
a ggplot2 plot
}
\description{
A function to draw bar (B) plot with point (P) and error (E) bars with t-test
}
