% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFC_multi.R
\name{plotFC_multi}
\alias{plotFC_multi}
\title{Density plot for flow cytometry}
\usage{
plotFC_multi(
  files,
  labels,
  channel,
  xlims,
  title = "xxx",
  fills,
  colors,
  alpha,
  fixed_ratio = 3
)
}
\arguments{
\item{files}{standard CSV files generated by NovoExpress software}

\item{labels}{labels}

\item{channel}{channel name in the CSV file}

\item{xlims}{two numeric values, specifying the lower limit and the upper limit of the scale}

\item{title}{plot title}

\item{fills}{colors used to fill the density}

\item{colors}{colors}

\item{alpha}{alpha}

\item{fixed_ratio}{ratio of the x axis and y axis}
}
\value{
a ggplot2 density plot
}
\description{
The function is designed to generate a multi-sample density plot for flow cytometry assay
}
\examples{
# none

}
