% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ_t.R
\name{plotQ_t}
\alias{plotQ_t}
\title{Gene expression plot for qPCR analysis}
\usage{
plotQ_t(
  file,
  genes,
  groups,
  title,
  fill,
  label.group,
  size.p = 2,
  size.text.x = 4,
  size.text.y = 3,
  vjust.p = -0.5,
  hjust.p = 0.5,
  angle.p = 0,
  scale.y = 1.3
)
}
\arguments{
\item{file}{a table file of gene expression generated by QuantStudio software}

\item{genes}{genes to plot}

\item{groups}{two groups}

\item{title}{plot title}

\item{fill}{colors}

\item{label.group}{group labels}

\item{size.p}{font size for p value}

\item{size.text.x}{font size for x-axis text}

\item{size.text.y}{font size for y-axis text}

\item{vjust.p}{vjust for p value}

\item{hjust.p}{hjust for p value}

\item{angle.p}{angle for p value}

\item{scale.y}{y scale}
}
\value{
a ggplot2 bar plot
}
\description{
Gene expression plot for qPCR analysis
}
